# HLG optimization for read_parquet + iloc