@TomAugspurger 2020-06-25
Question: In my head I was thinking of an approach like

1. Allow `iloc` in addition to getitem at https://github.com/dask/dask/blob/7138f470f0e55f2ebdb7638ddc4dfe2e78671403/dask/dataframe/optimize.py#L63-L65
2. For iloc, use `meta[block.indices[1][0]` at https://github.com/dask/dask/blob/7138f470f0e55f2ebdb7638ddc4dfe2e78671403/dask/dataframe/optimize.py#L75

Do you know if something like that has a chance of working?