@gforsyth 2020-06-25
> 1. Allow `iloc` in addition to getitem at

yeah, so I've found a better check which is checking that `block.output.startswith("iloc-")` -- there's no clean `operator` check since the `Blockwise` looks like 

```
> block.dsk
{'iloc-677f4093af82134a39609fb71cf9586a': (<function apply at 0x7fb6ff4808b0>, <function apply_and_enforce at 0x7fb701cfc940>, ['_0', '_1'], (<class 'dict'>, [['_func', <function iloc at 0x7fb701ccd790>], ['_meta', Empty DataFrame
Columns: [name, x]
Index: []]]))}
```

re 2: I think I got too fancy trying to handle slices -- might make more sense to take a first cut with single columns.