@gforsyth 2020-06-25
Ok, I think that's working (still as a separate optimization function).  I'll spend some time tomorrow to see what an overlay would look like.  Most of the code is duplicated, as you noted, just some special handling around column names vs. indices, I think.