@gforsyth 2020-06-26
Ok, this is ready for another look.  It was more straight-forward than I thought to combine them (woo!).  
Really the only bits where there is divergence with the existing `getitem` stuff is that the block indices are indices instead of column names and the check for an `iloc` doesn't have a nice `operator` check available.