@gforsyth 2020-06-26
> ```python
> # column order is A B C
> df = dd.read_parquet(...)[['B', 'A']]  # swap the columns
> result = df.iloc[:, 0]
> assert result.name == "B"
> assert result.compute().name == "B"
> ```

That's a great check, added as another test