@mrocklin 2020-06-26
Quick thought: what if instead we changed the definition of `df.iloc[:, ...]` to call `df[...]` in common case simple situations?

I understand that there are some tricky cases of iloc when we have duplicate column names, but my guess is that those are very rare.  By trying to centralize many API routes down to a few common operations we might be able to more easily reason about optimizations in the future.