@martindurant 2020-06-26
> what if instead we changed the definition of df.iloc[:, ...] to call df[...] in common case simple situations?

I think that's what I originally had conceived of. In that case, though, we would have to check for and forgo the optimisation in the case of duplicate labels. Note that parquet does *not* allow duplicates, but pyarrow will apply pandas names to columns on load, so duplicates are technically still possible.