@mrocklin 2020-06-26
Yeah, duplicate labels seem uncommon to me though.  I guess this comes down to figuring out the cost of maintaining an alternate iloc optimization code path vs the value of how often this code path will be used.  I don't have a good sense of that.