@gforsyth 2020-06-29
> Quick thought: what if instead we changed the definition of `df.iloc[:, ...]` to call `df[...]` in common case simple situations?

I'm game to try that out and see what it looks like.
I can think of two ways to handle this:
a) tweak the HLG and replace iloc with getitem there
b) change the `iloc` method on `dask.dataframe.core` to do a column name lookup and then dispatch to `__getitem__`

Any strong feelings on approach?