@mrocklin 2020-06-29
I would choose (b) personally.  I think that it's likely to be easier and
have broader reaching effects

On Mon, Jun 29, 2020 at 9:17 AM Gil Forsyth <notifications@github.com>
wrote:

> Quick thought: what if instead we changed the definition of df.iloc[:,
> ...] to call df[...] in common case simple situations?
>
> I'm game to try that out and see what it looks like.
> I can think of two ways to handle this:
> a) tweak the HLG and replace iloc with getitem there
> b) change the iloc method on dask.dataframe.core to do a column name
> lookup and then dispatch to __getitem__
>
> Any strong feelings on approach?
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/dask/dask/pull/6345#issuecomment-651221587>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AACKZTEANB3C33B3URQXVO3RZC5ABANCNFSM4OIQYZTQ>
> .
>