@TomAugspurger 2020-06-30
So this may not be needed after #6355?